backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
chromadb==0.4.24
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0
httpcore==0.18.0
numpy==1.24.3
onnxruntime==1.16.0
tokenizers==0.15.0
tqdm==4.66.1
posthog==3.0.2
pulsar-client==3.3.0
overrides==7.4.0
typing-extensions==4.8.0
